+++ capiq/tests/unit/test_capiq_client_gdsg.py
@@ -2,38 +2,10 @@ import unittest
 
 from mock import mock
 
-from capiq_client import CapIQClient
+from capiq.capiq_client import CapIQClient
 
 
-def mocked_gdsp_data_requests_post(*args, **kwargs):
-    class MockResponse:
-        def __init__(self, json_data, status_code):
-            self.json_data = json_data
-            self.status_code = status_code
-
-        def json(self):
-            return self.json_data
-
-    if args[0] is not None:
-        return MockResponse({"GDSSDKResponse": [{
-            "Headers": ["IQ_CLOSEPRICE"],
-            "Rows": [{"Row": ["46.80"]}],
-            "NumCols": 1,
-            "Seniority": "",
-            "Mnemonic": "IQ_CLOSEPRICE",
-            "Function": "GDSP",
-            "ErrMsg": None,
-            "Properties": {},
-            "StartDate": "",
-            "NumRows": 1,
-            "CacheExpiryTime": "0",
-            "SnapType": "",
-            "Frequency": "",
-            "Identifier": "TRIP:",
-            "Limit": ""
-        }]}, 200)
-
-def mocked_gdsp_no_data_requests_post(*args, **kwargs):
+def mocked_gdsg_data_requests_post(*args, **kwargs):
     class MockResponse:
         def __init__(self, json_data, status_code):
             self.json_data = json_data
@@ -47,6 +19,34 @@ def mocked_gdsp_no_data_requests_post(*args, **kwargs):
     elif args[0] == 'http://someotherurl.com/anothertest.json':
     """
     if args[0] is not None:
+        return MockResponse({"GDSSDKResponse": [      {
+            "Headers": ["IQ_CLOSEPRICE"],
+            "Rows": [{"Row": ["46.80"]}],
+            "NumCols": 1,
+            "Seniority": "",
+            "Mnemonic": "IQ_CLOSEPRICE",
+            "Function": "GDSG",
+            "ErrMsg": None,
+            "Properties": {},
+            "StartDate": "",
+            "NumRows": 1,
+            "CacheExpiryTime": "0",
+            "SnapType": "",
+            "Frequency": "",
+            "Identifier": "TRIP:",
+            "Limit": ""
+        }]}, 200)
+
+def mocked_gdsg_no_data_requests_post(*args, **kwargs):
+    class MockResponse:
+        def __init__(self, json_data, status_code):
+            self.json_data = json_data
+            self.status_code = status_code
+
+        def json(self):
+            return self.json_data
+
+    if args[0] is not None:
         return MockResponse(
             {
                 "GDSSDKResponse":
@@ -57,7 +57,7 @@ def mocked_gdsp_no_data_requests_post(*args, **kwargs):
                             "NumCols": 1,
                             "Seniority": "",
                             "Mnemonic": "IQ_CLOSEPRICE",
-                            "Function": "GDSP",
+                            "Function": "GDSG",
                             "ErrMsg": "SOME ERROR",
                             "Properties": {},
                             "StartDate": "",
@@ -71,16 +71,16 @@ def mocked_gdsp_no_data_requests_post(*args, **kwargs):
                     ]
             }, 200)
 
-class TestCapiqClientGdsp(unittest.TestCase):
+class TestCapiqClientGdsg(unittest.TestCase):
 
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdsp_data_requests_post)
-    def test_gdsp_data(self, mocked_post):
+    @mock.patch('capiq.capiq_client.requests.post', side_effect=mocked_gdsg_data_requests_post)
+    def test_gdsg_data(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdsp(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
+        return_value = ciq_client.gdsg(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], properties=[{}])
         self.assertEqual(return_value, {'TRIP:': {'close_price': '46.80'}})
 
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdsp_no_data_requests_post)
-    def test_gdsp_no_data(self, mocked_post):
+    @mock.patch('capiq.capiq_client.requests.post', side_effect=mocked_gdsg_no_data_requests_post)
+    def test_gdsg_no_data(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdsp(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
+        return_value = ciq_client.gdsg(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
         self.assertEqual(return_value, {'TRIP:': {'close_price': None}})
