+++ capiq/tests/unit/test_capiq_client_gdshe.py
@@ -2,10 +2,10 @@ import unittest
 
 from mock import mock
 
-from capiq_client import CapIQClient
+from capiq.capiq_client import CapIQClient
 
 
-def mocked_gdshv_data_requests_post(*args, **kwargs):
+def mocked_gdshe_data_requests_post(*args, **kwargs):
     class MockResponse:
         def __init__(self, json_data, status_code):
             self.json_data = json_data
@@ -20,7 +20,7 @@ def mocked_gdshv_data_requests_post(*args, **kwargs):
             "NumCols": 1,
             "Seniority": "",
             "Mnemonic": "IQ_CLOSEPRICE",
-            "Function": "GDSHV",
+            "Function": "GDSHE",
             "ErrMsg": None,
             "Properties": {},
             "StartDate": "",
@@ -32,7 +32,7 @@ def mocked_gdshv_data_requests_post(*args, **kwargs):
             "Limit": ""
         }]}, 200)
 
-def mocked_gdshv_no_data_requests_post(*args, **kwargs):
+def mocked_gdshe_no_data_requests_post(*args, **kwargs):
     class MockResponse:
         def __init__(self, json_data, status_code):
             self.json_data = json_data
@@ -52,7 +52,7 @@ def mocked_gdshv_no_data_requests_post(*args, **kwargs):
                             "NumCols": 1,
                             "Seniority": "",
                             "Mnemonic": "IQ_CLOSEPRICE",
-                            "Function": "GDSHV",
+                            "Function": "GDSHE",
                             "ErrMsg": "SOME ERROR",
                             "Properties": {},
                             "StartDate": "",
@@ -66,24 +66,24 @@ def mocked_gdshv_no_data_requests_post(*args, **kwargs):
                     ]
             }, 200)
 
-class TestCapiqClientGdshv(unittest.TestCase):
+class TestCapiqClientGdshe(unittest.TestCase):
 
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdshv_data_requests_post)
-    def test_gdshv_data(self, mocked_post):
+    @mock.patch('capiq.capiq_client.requests.post', side_effect=mocked_gdshe_data_requests_post)
+    def test_gdshe_data(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdshv(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], properties=[{}])
+        return_value = ciq_client.gdshe(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], properties=[{}])
         self.assertEqual(return_value, {'TRIP:': {'close_price': '46.80'}})
 
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdshv_no_data_requests_post)
-    def test_gdst_no_data(self, mocked_post):
+    @mock.patch('capiq.capiq_client.requests.post', side_effect=mocked_gdshe_no_data_requests_post)
+    def test_gdshe_no_data(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdshv(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
+        return_value = ciq_client.gdshe(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
         self.assertEqual(return_value, {'TRIP:': {'close_price': None}})
 
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdshv_data_requests_post)
+    @mock.patch('capiq.capiq_client.requests.post', side_effect=mocked_gdshe_data_requests_post)
     def test_gdst_data_no_properties(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdshv(
+        return_value = ciq_client.gdshe(
             ["TRIP"],
             ["IQ_CLOSEPRICE"],
             ["close_price"],
