+++ capiq/tests/unit/test_capiq_client_gdst.py
@@ -2,17 +2,21 @@ import unittest
 
 from mock import mock
 
-from capiq_client import CapIQClient
+from capiq.capiq_client import CapIQClient
 
 
-def mocked_gdshe_data_requests_post(*args, **kwargs):
+def mocked_gdst_data_requests_post(*args, **kwargs):
     class MockResponse:
         def __init__(self, json_data, status_code):
             self.json_data = json_data
             self.status_code = status_code
-
         def json(self):
             return self.json_data
+    """
+    if args[0] == 'http://someurl.com/test.json':
+        return MockResponse({"key1": "value1"}, 200)
+    elif args[0] == 'http://someotherurl.com/anothertest.json':
+    """
     if args[0] is not None:
         return MockResponse({"GDSSDKResponse": [      {
             "Headers": ["IQ_CLOSEPRICE"],
@@ -20,7 +24,7 @@ def mocked_gdshe_data_requests_post(*args, **kwargs):
             "NumCols": 1,
             "Seniority": "",
             "Mnemonic": "IQ_CLOSEPRICE",
-            "Function": "GDSHE",
+            "Function": "GDST",
             "ErrMsg": None,
             "Properties": {},
             "StartDate": "",
@@ -32,7 +36,7 @@ def mocked_gdshe_data_requests_post(*args, **kwargs):
             "Limit": ""
         }]}, 200)
 
-def mocked_gdshe_no_data_requests_post(*args, **kwargs):
+def mocked_gdst_no_data_requests_post(*args, **kwargs):
     class MockResponse:
         def __init__(self, json_data, status_code):
             self.json_data = json_data
@@ -40,7 +44,11 @@ def mocked_gdshe_no_data_requests_post(*args, **kwargs):
 
         def json(self):
             return self.json_data
-
+    """
+    if args[0] == 'http://someurl.com/test.json':
+        return MockResponse({"key1": "value1"}, 200)
+    elif args[0] == 'http://someotherurl.com/anothertest.json':
+    """
     if args[0] is not None:
         return MockResponse(
             {
@@ -52,7 +60,7 @@ def mocked_gdshe_no_data_requests_post(*args, **kwargs):
                             "NumCols": 1,
                             "Seniority": "",
                             "Mnemonic": "IQ_CLOSEPRICE",
-                            "Function": "GDSHE",
+                            "Function": "GDST",
                             "ErrMsg": "SOME ERROR",
                             "Properties": {},
                             "StartDate": "",
@@ -66,28 +74,29 @@ def mocked_gdshe_no_data_requests_post(*args, **kwargs):
                     ]
             }, 200)
 
-class TestCapiqClientGdshe(unittest.TestCase):
+class TestCapiqClientGdst(unittest.TestCase):
 
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdshe_data_requests_post)
-    def test_gdshe_data(self, mocked_post):
+    @mock.patch('capiq.capiq_client.requests.post', side_effect=mocked_gdst_data_requests_post)
+    def test_gdst_data(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdshe(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], properties=[{}])
+        return_value = ciq_client.gdst(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], properties=[{}])
         self.assertEqual(return_value, {'TRIP:': {'close_price': '46.80'}})
 
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdshe_no_data_requests_post)
-    def test_gdshe_no_data(self, mocked_post):
-        ciq_client = CapIQClient()
-        return_value = ciq_client.gdshe(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
-        self.assertEqual(return_value, {'TRIP:': {'close_price': None}})
-
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdshe_data_requests_post)
+    @mock.patch('capiq.capiq_client.requests.post', side_effect=mocked_gdst_data_requests_post)
     def test_gdst_data_no_properties(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdshe(
+        return_value = ciq_client.gdst(
             ["TRIP"],
             ["IQ_CLOSEPRICE"],
             ["close_price"],
             start_date="12/19/1980",
-            end_date="12/19/2000"
+            end_date="12/19/2000",
+            frequency="M"
         )
         self.assertEqual(return_value, {'TRIP:': {'close_price': '46.80'}})
+
+    @mock.patch('capiq.capiq_client.requests.post', side_effect=mocked_gdst_no_data_requests_post)
+    def test_gdst_no_data(self, mocked_post):
+        ciq_client = CapIQClient()
+        return_value = ciq_client.gdst(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
+        self.assertEqual(return_value, {'TRIP:': {'close_price': None}})
