+++ setup.py
@@ -1,0 +1,12 @@
+from distutils.core import setup
+
+setup(
+    name='capiq-python',
+    version='0.1',
+    packages=['capiq', 'capiq.tests', 'capiq.tests.unit'],
+    url='',
+    license='',
+    author='gmann',
+    author_email='',
+    description=''
+)
